Move reusable parts into GeneratorFromFileBase.
