+++ Lombiq.HelpfulLibraries.SourceGenerators/GeneratorFromFileBase.cs
@@ -1,0 +1,112 @@
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using System;
+using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Linq;
+
+namespace Lombiq.HelpfulLibraries.SourceGenerators;
+
+public abstract class GeneratorFromFileBase : IIncrementalGenerator
+{
+    protected abstract Type Attribute { get; }
+
+    /// <summary>
+    /// Generate code action. It will be executed on specific nodes (<see cref="ClassDeclarationSyntax"/> annotated with
+    /// the attribute).
+    /// </summary>
+    /// <param name="context">Source generation context used to add source files.</param>
+    /// <param name="compilation">Compilation used to provide access to the Semantic Model.</param>
+    /// <param name="classDeclarations">Nodes annotated with the attribute that trigger the generate action.</param>
+    protected abstract void GenerateCode(
+        SourceProductionContext context,
+        Compilation compilation,
+        ImmutableArray<(ClassDeclarationSyntax Syntax, List<Dictionary<string, string>> Dictionary)> classDeclarations,
+        IDictionary<string, string> additionalFiles);
+
+    public void Initialize(IncrementalGeneratorInitializationContext context)
+    {
+        var fullName = $"{typeof(Attribute).Namespace}.{nameof(Attribute.Name)}";
+
+        // Filter classes annotated with the attribute.
+        // Only filtered Syntax Nodes can trigger code generation.
+        var provider = context.SyntaxProvider
+            .CreateSyntaxProvider(
+                (node, _) => node is ClassDeclarationSyntax,
+                (syntaxContext, _) => GetClassDeclarationForSourceGen(syntaxContext, fullName))
+            .Where(tuple => tuple.ReportAttributeFound)
+            .Select((tuple, _) => (tuple.Syntax, tuple.AttributesData));
+
+        var additionalFiles = context.AdditionalTextsProvider
+            .Where(static file => file.Path.EndsWith(".json", StringComparison.OrdinalIgnoreCase));
+
+        var namesAndContents = additionalFiles
+            .Select((file, cancellationToken) =>
+                (Content: file.GetText(cancellationToken)?.ToString(),
+                    file.Path));
+
+        var filesAndContents = new Dictionary<string, string>();
+
+        context.RegisterSourceOutput(namesAndContents.Collect(), (_, contents) =>
+        {
+            foreach ((string? content, string path) in contents)
+            {
+                // Check if path already exists
+                if (!filesAndContents.ContainsKey(path))
+                {
+                    // Add to the dictionary
+                    filesAndContents.Add(path, content ?? string.Empty);
+                }
+            }
+        });
+
+        // Generate the source code.
+        context.RegisterSourceOutput(
+            context.CompilationProvider.Combine(provider.Collect()),
+            (productionContext, tuple) => GenerateCode(productionContext, tuple.Left, tuple.Right, filesAndContents));
+    }
+
+    /// <summary>
+    /// Checks whether the Node is annotated with the <see cref="Attribute"/> and maps syntax context to
+    /// the specific node type (<see cref="ClassDeclarationSyntax"/>).
+    /// </summary>
+    /// <param name="context">Syntax context, based on <see cref="SyntaxValueProvider.CreateSyntaxProvider"/> predicate.</param>
+    /// <param name="fullName">The namespace and type name of the target attribute.</param>
+    /// <returns>The specific cast and whether the attribute was found.</returns>
+    private static (ClassDeclarationSyntax Syntax, bool ReportAttributeFound, List<Dictionary<string, string>> AttributesData)
+        GetClassDeclarationForSourceGen(GeneratorSyntaxContext context, string fullName)
+    {
+        var classDeclarationSyntax = (ClassDeclarationSyntax)context.Node;
+        var attributesData = classDeclarationSyntax.AttributeLists
+            .SelectMany(list => list.Attributes)
+            .Select(attributeSyntax => GetAttributeArguments(context, attributeSyntax, fullName))
+            .OfType<Dictionary<string, string>>().ToList();
+
+        return (classDeclarationSyntax, attributesData.Count > 0, attributesData);
+    }
+
+    private static Dictionary<string, string>? GetAttributeArguments(
+        GeneratorSyntaxContext context,
+        AttributeSyntax attributeSyntax,
+        string fullName)
+    {
+        // Skip if we can't get the symbol or if the full name doesn't match.
+        if (context.SemanticModel.GetSymbolInfo(attributeSyntax).Symbol is not IMethodSymbol attributeSymbol ||
+            attributeSymbol.ContainingType.ToDisplayString() != fullName)
+        {
+            return null;
+        }
+
+        var arguments = attributeSyntax.ArgumentList?.Arguments
+            .Select(argument => argument.Expression)
+            .OfType<LiteralExpressionSyntax>()
+            .Select((literalExpression, index) => new
+            {
+                Key = attributeSymbol.Parameters[index].Name,
+                Value = literalExpression.Token.Text,
+            })
+            .ToDictionary(keyValuePair => keyValuePair.Key, keyValuePair => keyValuePair.Value) ?? [];
+
+        return arguments;
+    }
+}
